Fix bug: messages out of order.
According to the issue, the messages are to display in chronological order from top to bottom.  Through initial testing I noticed that they seemed to simply be displaying in reverse order so I just made a shallow copy of the data and reversed it.  Note that simply calling reverse() on the original prop messages would be illegal modification of the state because reverse works in place, modifying the array it is called on, not returning a new array.