Fix bug: Direct mutation of state
New messages were not rendered upon submission by the sender because no action was dispatched.  When state is updated properly, by dispatching actions, the UI components that subscribe to that state (via connect mapStateToProps ) are notified of the change. When state is mutated directly with object assignment the notification process is bypassed.