Move choice of Link down from Layout to Navigation
Responsive Layout was importing two different Navigation Links and using a prop "page" to determine which one to render.  Though it functioned correctly, it was confusing because nothing about the page name "Responsive Layout" suggests this file would have that responsibility.  Now instead it imports one Link and passes along the page prop.  The Navigation file is still using that prop to choose between the same two components, but now that choice is happening closer to where those two links are defined.  This organizational change would make this logic much easier to locate for someone new to this codebase.